backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── AWS SDK ───────────────────────────────────────────────────────────────────
boto3==1.34.84
botocore==1.34.84

# ── AI ────────────────────────────────────────────────────────────────────────
google-generativeai==0.5.4

# ── Auth ──────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
aiosqlite==0.20.0

# ── Data Validation ───────────────────────────────────────────────────────────
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1

# ── Network (optional — uncomment to enable live packet capture) ──────────────
# scapy==2.5.0

# ── Development / Testing (install separately) ────────────────────────────────
# pip install pytest pytest-asyncio httpx
